// WADNR.Web/src/app/shared/generated/model/organization-type-grid-row.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class OrganizationTypeGridRow { 
    OrganizationTypeID?: number;
    OrganizationTypeName?: string | null;
    OrganizationTypeAbbreviation?: string | null;
    LegendColor?: string | null;
    ShowOnProjectMaps?: boolean;
    IsDefaultOrganizationType?: boolean;
    IsFundingType?: boolean;
    OrganizationCount?: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface OrganizationTypeGridRowForm { 
    OrganizationTypeID?: FormControl<number>;
    OrganizationTypeName?: FormControl<string>;
    OrganizationTypeAbbreviation?: FormControl<string>;
    LegendColor?: FormControl<string>;
    ShowOnProjectMaps?: FormControl<boolean>;
    IsDefaultOrganizationType?: FormControl<boolean>;
    IsFundingType?: FormControl<boolean>;
    OrganizationCount?: FormControl<number>;
}

export class OrganizationTypeGridRowFormControls { 
    public static OrganizationTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static OrganizationTypeName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static OrganizationTypeAbbreviation = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static LegendColor = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ShowOnProjectMaps = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static IsDefaultOrganizationType = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static IsFundingType = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static OrganizationCount = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}